use Python 3.9

+++ Lab1/scripts/deployment.sh
@@ -49,11 +49,22 @@ if [[ $server -eq 1 ]]; then
   cd ../../scripts
 
   echo "copying profile to labs"
-  cp ../../Lab2/scripts/.env.template ../../Lab2/scripts/.env
-  cp ../../Lab3/scripts/.env.template ../../Lab3/scripts/.env
-  cp ../../Lab4/scripts/.env.template ../../Lab4/scripts/.env
-  cp ../../Lab5/scripts/.env.template ../../Lab5/scripts/.env
-  cp ../../Lab6/scripts/.env.template ../../Lab6/scripts/.env
+  if ! test -f ../../Lab2/scripts/.env; then
+    cp ../../Lab2/scripts/.env.template ../../Lab2/scripts/.env
+  fi
+  if ! test -f ../../Lab3/scripts/.env; then
+    cp ../../Lab3/scripts/.env.template ../../Lab3/scripts/.env
+
+  fi
+  if ! test -f ../../Lab4/scripts/.env; then
+    cp ../../Lab4/scripts/.env.template ../../Lab4/scripts/.env
+  fi
+  if ! test -f ../../Lab5/scripts/.env; then
+    cp ../../Lab5/scripts/.env.template ../../Lab5/scripts/.env
+  fi
+  if ! test -f ../../Lab6/scripts/.env; then
+    cp ../../Lab6/scripts/.env.template ../../Lab6/scripts/.env
+  fi
   ex -sc '%s/AWS_PROFILE=.*/AWS_PROFILE=\"'$AWS_PROFILE'\"/|x' ../../Lab2/scripts/.env
   ex -sc '%s/AWS_PROFILE=.*/AWS_PROFILE=\"'$AWS_PROFILE'\"/|x' ../../Lab3/scripts/.env
   ex -sc '%s/AWS_PROFILE=.*/AWS_PROFILE=\"'$AWS_PROFILE'\"/|x' ../../Lab4/scripts/.env
